requirements.txt
pytest==8.4.2
selenium==4.38.0
allure-pytest==2.14.1
requests==2.32.5